app/requirements.txt
APScheduler==3.6.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
decorator==4.4.2
dnspython==2.0.0
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
pycparser==2.20
pymongo==3.11.1
python-dotenv==0.15.0
python-telegram-bot==13.1
python-twitter==3.5
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
